requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
qrcode==7.4.2
Pillow==12.1.1
reportlab==4.0.9
gunicorn==21.2.0
psycopg2-binary==2.9.9
